requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
opencv-python
Pillow
scikit-learn
